Add unit tests.

+++ src/main/java/eu/tax_rate/app/database/model/TaxRateDocument.java
@@ -32,5 +32,5 @@ public class TaxRateDocument {
     String taxRate;
 
     /** Priority of this tax rate configuration, to be used when multiple tax rates are found */
-    int priority;
+    long priority;
 }
